# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tasing API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Tasing API Server"
        }
    ],
    "paths": {
...
        "/api/dashboard/purchases": {
            "get": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Display a listing of the user's purchase files",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "project",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Purchase files retrived successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                },
                "security": [
                    {
                        "": []
                    }
                ]
            }
        },
        "/api/dashboard/purchases/{purchaseCode}/records": {
            "get": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Display the specified user's purchase",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "purchaseCode",
                        "in": "path",
                        "description": "code of purchase",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Purchase file retrived successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Access Denied."
                    },
                    "404": {
                        "description": "Purchase File not found."
                    }
                },
                "security": [
                    {
                        "": []
                    }
                ]
            }
        },
        "/api/login/{provider}": {
            "get": {
                "tags": [
                    "OAuth"
                ],
                "summary": "Retrieve the url to redirect to the given provider authentication page.",
                "operationId": "redirect",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Provider redirect url retrieve successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login/{provider}/callback": {
            "get": {
                "tags": [
                    "OAuth"
                ],
                "summary": "Obtain the user information from given provider, retrieve acces token.",
                "operationId": "callback",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User logged successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/User"
                                                },
                                                "access_token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "description": "name",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "lastname",
                        "type": "string"
                    },
                    "email": {
                        "description": "email",
                        "type": "string"
                    },
                    "phone_number1": {
                        "description": "phone_number1",
                        "type": "string"
                    },
                    "phone_number2": {
                        "description": "phone_number2",
                        "type": "string"
                    },
                    "country_code": {
                        "description": "country_code",
                        "type": "string"
                    },
                    "address_line1": {
                        "description": "address_line1",
                        "type": "string"
                    },
                    "address_line2": {
                        "description": "address_line2",
                        "type": "string"
                    },
                    "city": {
                        "description": "city",
                        "type": "string"
                    },
                    "region": {
                        "description": "region",
                        "type": "string"
                    },
                    "zipcode": {
                        "description": "zipcode",
                        "type": "string"
                    },
                    "company_name": {
                        "description": "company_name",
                        "type": "string"
                    },
                    "company_number": {
                        "description": "company_number",
                        "type": "string"
                    },
                    "email_verified_at": {
                        "description": "email_verified_at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}